% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_weighting.R
\name{assess_weighting}
\alias{assess_weighting}
\title{Assess Weighting}
\usage{
assess_weighting(targets, data)
}
\arguments{
\item{targets}{A list of weighting targets (see the gen_us() function to generate these easily)}

\item{data}{Data for which you want to generate weights.}
}
\description{
Takes your target weighting and your data and returns data and plots compared your observed %ages with their desired targets. This can be used to help select what variables to include and determine the degree of deviation from the targets in the sample.
}
