% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_corr_post.R
\name{bayes_corr_post}
\alias{bayes_corr_post}
\title{Bayesian correlation posterior}
\usage{
bayes_corr_post(post)
}
\arguments{
\item{post}{A draw from a posterior of the brms rescor posterior object}
}
\description{
Generate the posterior for residual/pairwise correlations, the inverted correlation matrix, and partial correlation matrix.
}
