% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_mrp.R
\name{beta_mrp}
\alias{beta_mrp}
\title{Fit Beta MRP Model}
\usage{
beta_mrp(
  variable_name,
  variable_label = NULL,
  convert_beta = TRUE,
  my_data = set_my_data,
  save_model = save_my_model,
  save_epred = save_my_epred,
  my_draw_ids = seq(from = 1, to = 4000, length.out = 1000) \%>\% round(),
  my_prior = c(set_prior("normal(0 , 1)", class = "Intercept"),
    set_prior("normal(0 , .67)", class = "b"), set_prior("exponential(2)", class = "sd"),
    set_prior("normal(5 , 2.5)", class = "phi", lb = 0.01)),
  my_init = 0,
  my_refresh = 250,
  my_iter = set_my_iter,
  my_warmup = set_my_warmup,
  my_poststrat = set_my_poststrat,
  my_adapt_delta = set_my_adapt_delta,
  mrp_form = NULL,
  name_addition = ""
)
}
\arguments{
\item{variable_name}{String. The name of the continuous variable to be modeled.}

\item{variable_label}{Optional. A human-readable label for the variable. Defaults to \code{variable_name}.}

\item{convert_beta}{Logical. Whether to transform the outcome variable to be suitable for Beta regression. Defaults to \code{TRUE}.}

\item{my_data}{A data frame containing the data for the regression model. Defaults to \code{set_my_data}.}

\item{save_model}{Logical. If \code{TRUE}, saves the fitted model object. Defaults to \code{save_my_model}.}

\item{save_epred}{Logical. If \code{TRUE}, saves the posterior predictive draws. Defaults to \code{save_my_epred}.}

\item{my_draw_ids}{Numeric vector. Indices of posterior draws to use for predictions. Defaults to a sequence of 1000 evenly spaced draws from 4000 iterations.}

\item{my_prior}{A list of prior distributions for the model. Defaults to:
\itemize{
  \item \code{"normal(0 , 1)"} for the intercept.
  \item \code{"normal(0 , 0.67)"} for the coefficients.
  \item \code{"exponential(2)"} for random effects.
  \item \code{"normal(5 , 2.5)"} for the precision parameter \code{phi}, with a lower bound of \code{0.01}.
}}

\item{my_init}{Initial values for the model. Defaults to \code{0}.}

\item{my_refresh}{Numeric. Frequency of progress updates during model fitting. Defaults to \code{250}.}

\item{my_iter}{Numeric. Total number of iterations per chain. Defaults to \code{set_my_iter}.}

\item{my_warmup}{Numeric. Number of warmup iterations for the sampler. Defaults to \code{set_my_warmup}.}

\item{my_poststrat}{A data frame for poststratification. Defaults to \code{set_my_poststrat}.}

\item{my_adapt_delta}{Numeric. The Stan sampler control parameter to improve convergence. Defaults to \code{set_my_adapt_delta}.}

\item{mrp_form}{Optional string specifying a custom formula for the regression model. If \code{NULL}, a default formula is used.}
}
\value{
A list containing:
  \itemize{
    \item \code{"model"}: The fitted \code{brms} model object.
    \item \code{"epred"}: Posterior epred from the fitted model.
    \item \code{"phi"}: Posterior draws of the precision parameter \code{phi}.
  }
}
\description{
Perform multilevel regression for a continuous outcome variable using a Beta distribution.
}
\details{
This function fits a Beta regression model with multilevel structure using the \code{brms} package and \code{cmdstanr} backend.
If \code{convert_beta} is set to \code{TRUE}, the outcome variable is transformed to fit within the (0, 1) range using \code{jimbilben::transform_beta}.
The function constructs a model formula based on a default or user-specified structure, applies user-defined priors, and fits the model.
Posterior predictions and the precision parameter \code{phi} are returned, and optionally saved for future use.
}
