% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_mrp.R
\name{binary_mrp}
\alias{binary_mrp}
\title{Run Binary MRP}
\usage{
binary_mrp(
  variable_name,
  variable_label = NULL,
  my_data = set_my_data,
  save_model = save_my_model,
  save_epred = save_my_epred,
  my_prior = c(set_prior("normal(0 , 2)", class = "Intercept"),
    set_prior("normal(0 , 1)", class = "b"), set_prior("exponential(2)", class = "sd")),
  my_init = 0,
  my_warmup = set_my_warmup,
  my_refresh = 250,
  my_iter = set_my_iter,
  my_poststrat = set_my_poststrat,
  my_adapt_delta = set_my_adapt_delta,
  weighted = FALSE,
  weight_name = "mod_weights_mrp",
  mrp_form = NULL
)
}
\arguments{
\item{variable_name}{String. The name of the binary variable to be modeled.}

\item{variable_label}{Optional. A human-readable label for the variable. Defaults to \code{variable_name}.}

\item{my_data}{A data frame containing the data for the regression model. Defaults to \code{set_my_data}.}

\item{save_model}{Logical. If \code{TRUE}, saves the fitted model object to an mrp_models folder in your directory. Defaults to \code{save_my_model}.}

\item{save_epred}{Logical. If \code{TRUE}, saves the posterior epred draws to an mrp_epreds folder in your directory. Defaults to \code{save_my_epred}.}

\item{my_prior}{A list of prior distributions for the model. Defaults to:
\itemize{
  \item \code{Intercept: normal(0, 2)}
  \item \code{Coefficients: normal(0, 1)}
  \item \code{Random effects: exponential(2)}
}}

\item{my_init}{Initial values for the model. Default is \code{0}.}

\item{my_warmup}{Number of warmup iterations for the sampler. Default is \code{set_my_warmup}.}

\item{my_refresh}{Frequency of progress updates during model fitting. Default is \code{250}.}

\item{my_iter}{Total number of iterations per chain. Default is \code{set_my_iter}.}

\item{my_poststrat}{A data frame for poststratification. Default is \code{set_my_poststrat}.}

\item{my_adapt_delta}{Numeric value for the Stan sampler control parameter to improve convergence. Default is \code{set_my_adapt_delta}.}

\item{weighted}{Logical. If \code{TRUE}, performs weighted regression using the specified weights. Defaults to \code{FALSE}.}

\item{weight_name}{The name of the weight variable in \code{my_data}. Defaults to \code{"mod_weights_mrp"}.}

\item{mrp_form}{Optional string specifying a custom formula for the regression model. If \code{NULL}, a default formula is used.}
}
\value{
A list containing:
  \itemize{
    \item \code{"model"}: The fitted \code{brms} model object.
    \item \code{"epred"}: Posterior predictions from the fitted model.
  }
}
\description{
Perform the regression and epred steps of a multilevel regression and poststratification (MRP) for a binary outcome variable.
}
\details{
This function performs multilevel Bayesian regression using the \code{brms} package.
It constructs a model formula based on the specified or default structure, fits the model
using the \code{cmdstanr} backend, and generates posterior epred draws for
a target population. The model and predictions can be optionally saved for future use.
}
\note{
Requires the \code{brms} and \code{cmdstanr} packages.
Ensure that the \code{my_data} and \code{my_poststrat} arguments are properly formatted.
}
