% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_label.R
\name{break_label}
\alias{break_label}
\title{Break label}
\usage{
break_label(data, from = label, to = label_break)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{from}{The variable name of the original label - defaults to label}

\item{to}{The variable name for the converted label - defaluts to label_break}
}
\description{
Add a line break to a label of the form 'X.XX [Y.YY - Z.ZZ]' so that the confidence interval comes on a new line
}
