% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmrp.R
\name{brmrp}
\alias{brmrp}
\title{Rapid MRP with brms}
\usage{
brmrp(
  data,
  variable = "outcome",
 
    pred_formula = "~ (1 | state) + (1 | race) + (1 | age_fine) +\\n                                 (1 | education_collapse) + (1 | income_ces) + male + (1 | male:race) +\\n                                 (1 | education_collapse:age_fine) + (1 | partyid) + factor(region) + z_repvote",
  .prior,
  .family = brms::categorical(),
  .iter = 4000,
  .warmup = 1000,
  zero_init = TRUE,
  .chains = 4,
  .cores = 4,
  .control = list(adapt_delta = 0.995, max_treedepth = 15),
  .stan_model_args = list(stanc_options = list("O1")),
  .threading = 8,
  save_to_file = TRUE,
  .seed = 1010
)
}
\arguments{
\item{data}{The data frame containing the variables used in the MRP model}

\item{variable}{The name of the outcome variable (Default: "outcome")}

\item{pred_formula}{The predictor formula for the MRP model (Default: "~ (1 | state) + (1 | race) + (1 | age_fine) +
(1 | education_collapse) + (1 | income_ces) + male + (1 | male:race) +
(1 | education_collapse:age_fine) + (1 | partyid) + factor(region) + z_repvote")}

\item{.prior}{A character vector of prior specifications compatible with the brms package}

\item{.family}{The family of the outcome variable (Default: brms::categorical())}

\item{.iter}{The number of iterations for the MCMC sampler (Default: 4000)}

\item{.warmup}{The number of warmup iterations (Default: 1000)}

\item{zero_init}{A logical value indicating whether to initialize parameters at zero (Default: TRUE)}

\item{.chains}{The number of chains for the MCMC sampler (Default: 4)}

\item{.cores}{The number of cores used for parallel computation (Default: 4)}

\item{.control}{A list of control arguments for the MCMC sampler (Default: list(adapt_delta = .995, max_treedepth = 15))}

\item{.stan_model_args}{A list of arguments to be passed to the Stan model (Default: list(stanc_options = list('O1')))}

\item{.threading}{The number of threads for within-chain parallelization (Default: 8)}

\item{save_to_file}{A logical value indicating whether to save the model to a file (Default: TRUE)}

\item{.seed}{The random seed for reproducibility (Default: 1010)}
}
\value{
A brmsfit object containing the results of the MRP model
}
\description{
A convenience function to run Multilevel Regression and Poststratification (MRP) using the brms package with reasonable default settings.
}
\examples{
# Example with a data frame containing MRP variables
# data <- read.csv("path/to/your/data.csv")
# priors <- mrp_cat_prior(data$region)
# brmrp(data, variable = "outcome", .prior = priors)
}
