% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_mrp_fit.R
\name{categorical_mrp_fit}
\alias{categorical_mrp_fit}
\title{Fit Categorical MRP Model}
\usage{
categorical_mrp_fit(
  variable_name,
  my_data = set_my_data,
  variable_label = NULL,
  reference_level = "Neither agree nor disagree",
  my_intercept_prior = "normal(0, 2)",
  my_b_prior = "normal(0, 1)",
  my_sd_prior = "exponential(3)",
  save_model = save_my_model,
  my_init = 0,
  my_refresh = 100,
  my_iter = set_my_iter,
  my_warmup = set_my_warmup,
  my_adapt_delta = 0.99,
  mrp_form = NULL,
  name_addition = ""
)
}
\arguments{
\item{variable_name}{String. The name of the categorical variable to be modeled.}

\item{my_data}{A data frame containing the data for the regression model. Defaults to \code{my_data}.}

\item{variable_label}{Optional. A human-readable label for the variable. Defaults to \code{variable_name}.}

\item{reference_level}{String. The reference level for the categorical variable. Defaults to \code{"Neither agree nor disagree"}.}

\item{my_intercept_prior}{String. Prior for the intercept. Defaults to \code{"normal(0, 2)"}.}

\item{my_b_prior}{String. Prior for the coefficients. Defaults to \code{"normal(0, 1)"}.}

\item{my_sd_prior}{String. Prior for the standard deviation of random effects. Defaults to \code{"exponential(3)"}.}

\item{save_model}{Logical. If \code{TRUE}, saves the fitted model object. Defaults to \code{save_my_model}.}

\item{my_init}{Initial values for the model. Defaults to \code{0}.}

\item{my_refresh}{Numeric. Frequency of progress updates during model fitting. Defaults to \code{100}.}

\item{my_iter}{Numeric. Total number of iterations per chain. Defaults to \code{set_my_iter}.}

\item{my_warmup}{Numeric. Number of warmup iterations for the sampler. Defaults to \code{set_my_warmup}.}

\item{my_adapt_delta}{Numeric. The Stan sampler control parameter to improve convergence. Defaults to \code{0.99}.}

\item{mrp_form}{Optional string specifying a custom formula for the regression model. If \code{NULL}, a default formula is used.}

\item{name_addition}{String, defaults to "" (i.e., nothing). Can provide a string to more uniquely identify what the file will be named as.}
}
\value{
A list containing:
  \itemize{
    \item \code{"model"}: The fitted \code{brms} model object.
  }
}
\description{
Perform multilevel regression for a categorical outcome variable using the \code{brms} package.
}
\details{
This function fits a categorical multilevel regression model using \code{brms} with \code{cmdstanr} as the backend.
It applies specified priors to the intercept, coefficients, and random effects using \code{mrp_cat_prior}.
The function constructs a default or user-specified model formula, adjusts the categorical variable's reference level, and fits the model.
Optionally, the fitted model can be saved for future use.
}
