% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_xy.R
\name{circle_xy}
\alias{circle_xy}
\title{Circle XY}
\usage{
circle_xy(degrees = 0, radius = 1, center_x = 0, center_y = 0)
}
\arguments{
\item{degrees}{The angle (or angles) around the circle that you wish to compute the x y coordinates from.}

\item{radius}{The radius of the circle to compute.}

\item{center_x}{The center of the circle in the y plane.}
}
\description{
Get the x and y coordinates for points around a circle, with a given center, radius, and degree.
}
