% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_expand.R
\name{code_expand}
\alias{code_expand}
\title{Expand code}
\usage{
code_expand(type = "mult", low = 0, high = 0.05)
}
\arguments{
\item{type}{String, defaults to "mult" (vs. "add")}

\item{low}{Numeric, expansion at the bottom of the axis - defaults to 0}

\item{high}{Numeric, expansion at the top of the axis - defaults to .05}
}
\description{
Print out code for ggplot axis expansion
}
