% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_percent.R
\name{code_percent}
\alias{code_percent}
\title{Percent code}
\usage{
code_percent(axis = "x")
}
\arguments{
\item{axis}{String, the axis being referenced - defaults to "x" (vs. "y")}
}
\description{
Print out code for ggplot axis expansion
}
