% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_psup_steps.R
\name{code_psup_steps}
\alias{code_psup_steps}
\title{PSup steps code}
\usage{
code_psup_steps(item_prefix = "item")
}
\description{
Print out code for the steps to conduct Bayesian within subjects PSup
}
