% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_rect.R
\name{code_rect}
\alias{code_rect}
\title{Rectangle code}
\usage{
code_rect(
  min = 100,
  max = 125,
  fill = "grey99",
  color = "grey95",
  linewidth = 0.33
)
}
\arguments{
\item{axis}{String, the axis being referenced - defaults to "x" (vs. "y")}
}
\description{
Print out code for ggplot rectangle and label text
}
