% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_summarise.R
\name{code_summarise}
\alias{code_summarise}
\title{Print summarise code}
\usage{
code_summarise(outcome = "rating")
}
\arguments{
\item{outcome}{The name of the variable that is being summarised}
}
\description{
Print out code for a commonly desired summary and corresponding label
}
