% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_psup_within.R
\name{conduct_psup_within}
\alias{conduct_psup_within}
\title{Conduct within subjects PSup}
\usage{
conduct_psup_within(
  variable,
  data,
  percent = FALSE,
  decimals = 2,
  prior = NULL,
  prefix = "psup_",
  sep = "_min_",
  intercept_prior = "normal(0 , 1.25)",
  dpar_prior = "normal(0 , 1)",
  my_iter = 2000,
  my_warmup = 500,
  weighted = FALSE,
  weight_name = "mod_weights"
)
}
\arguments{
\item{variable}{The name of the psup comparison variable}

\item{data}{The data frame where the variable is}

\item{percent}{Whether or not to show the PSup as a percentage format or not (defaults to FALSE)}

\item{decimals}{How many decimals to show in the summary label - defaults to 2}

\item{prior}{A prior for the analysis}

\item{prefix}{The prefix string, if any, in the psup comparison column - defaults to "psup_".}

\item{sep}{The separator string, if any, in the psup comparison column - defaults to "_min_".}
}
\value{
A brms fit object and several posterior objects and summaries
}
\description{
Performs a Bayesian within subjects probability of superiority analysis
}
