% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_excel_date.R
\name{convert_excel_date}
\alias{convert_excel_date}
\title{Convert Excel Date to POSIXct}
\usage{
convert_excel_date(excel_date)
}
\arguments{
\item{excel_date}{A numeric value representing the Excel date-time number to be converted.}
}
\value{
A POSIXct object representing the date and time.
}
\description{
Converts an Excel date-time number to a POSIXct date-time object. Excel stores dates as serial numbers
where the integer part represents the date and the fractional part represents the time. This function
converts such a number into a standard R date-time format.
}
\details{
The function first converts the input to a numeric value to ensure compatibility.
The origin for Excel dates (December 30, 1899) is used to convert the integer part to a date.
The fractional part of the number is then converted to seconds and added to the date to get
the full date-time. The function assumes the date-time is in the UTC timezone.
}
\examples{
convert_excel_date(44204.75) # Returns the POSIXct date-time for the Excel number 44204.75

}
