% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_kmeans.R
\name{do_kmeans}
\alias{do_kmeans}
\title{Do KMeans}
\usage{
do_kmeans(nclusters = 3, data_with_id, id_name = "id", decimals = 1)
}
\arguments{
\item{nclusters}{Numeric, indicating the number of clusters to generate (use NbClust::NbClust() to determine how many clusters are best)}

\item{data_with_id}{Your data to be clustered. It should include a unique id column and the target columns to be clustered on. Rows containing missing values will be removed.}

\item{id_name}{A string indicating the name of your unique id column, defaults to "id"}

\item{decimals}{Numeric, indicating the number of decimal places to display in the output graphs. Defaults to 1}
}
\description{
Performs a kmeans clustering analysis, with plots that show cluster characteristics, as well as returning data linking ids to cluster numbers.
}
