% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pca.R
\name{do_pca}
\alias{do_pca}
\title{Do Principal Component Analysis (PCA)}
\usage{
do_pca(data, id = "rowid", .center = TRUE, .scale = TRUE, limit_comps = NULL)
}
\arguments{
\item{data}{A data frame or tibble containing the data for PCA. The dataset should
include a column specified by the `id` argument that will be excluded from the analysis.}

\item{id}{A character string specifying the name of the column to be excluded from the PCA.
This column typically contains row identifiers. Default is "rowid".}

\item{.center}{A logical value indicating whether the variables should be centered before
PCA is performed. Default is TRUE.}

\item{.scale}{A logical value indicating whether the variables should be scaled before
PCA is performed. Default is TRUE.}

\item{limit_comps}{An optional integer specifying the number of principal components to
limit in the loadings plot. If NULL, all components are included. Default is NULL.}
}
\value{
A list containing the following elements:
  \item{pca}{The PCA object resulting from `prcomp`.}
  \item{individual_scores}{A tibble of individual scores (principal component scores) with
  the specified ID column included.}
  \item{scree}{A ggplot object representing the scree plot of eigenvalues.}
  \item{loadings_plot}{A ggplot object representing the loadings plot of the PCA components.}
  \item{plot_1}{A ggplot object representing the PCA individuals plot, colored by cos2 values.}
  \item{plot_2}{A ggplot object representing the PCA variables plot, colored by contributions.}
  \item{plot_3}{A ggplot object representing the PCA biplot.}
  \item{loadings}{A tibble of PCA loadings with the variable names and their respective loadings.}
  \item{var_diagnostics}{A list of diagnostics related to the PCA variables from `factoextra::get_pca_var`.}
  \item{ind_diagnostics}{A list of diagnostics related to the PCA individuals from `factoextra::get_pca_ind`.}
}
\description{
This function performs Principal Component Analysis (PCA) on a given dataset,
excluding the specified ID column. It also generates various plots and diagnostics
related to the PCA, including scree plots, biplots, and loading plots.
}
\examples{
# Example usage of the do_pca function
data <- mtcars
result <- do_pca(data, id = "carb")
result$scree
result$plot_1

}
