% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_age.R
\name{factor_age}
\alias{factor_age}
\title{Factor Age}
\usage{
factor_age(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
