% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_age_fine.R
\name{factor_age_fine}
\alias{factor_age_fine}
\title{Factor Age Fine}
\usage{
factor_age_fine(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
