% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_education.R
\name{factor_education}
\alias{factor_education}
\title{Factor Education}
\usage{
factor_education(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
