% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_education_collapse.R
\name{factor_education_collapse}
\alias{factor_education_collapse}
\title{Factor Education Collapse}
\usage{
factor_education_collapse(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
