% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_income.R
\name{factor_income}
\alias{factor_income}
\title{Factor Income}
\usage{
factor_income(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
