% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_male.R
\name{factor_male}
\alias{factor_male}
\title{Factor Male}
\usage{
factor_male(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
