% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_partyid.R
\name{factor_partyid}
\alias{factor_partyid}
\title{Factor PartyID}
\usage{
factor_partyid(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
