% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_race.R
\name{factor_race}
\alias{factor_race}
\title{Factor Race}
\usage{
factor_race(data)
}
\arguments{
\item{data}{The data to be factored}
}
\description{
factor_xyz functions correctly order demographic variables used in weighting and poststratification and, where necessary, shorten their names.
}
