% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_us_2022.R
\name{gen_us_2022}
\alias{gen_us_2022}
\title{gen_us_22}
\usage{
gen_us_2022(
  target_variable = "options",
  .moving_average = 1,
  .gallup_url = "insert your gallup URL",
  .stop_date = NULL,
  partyid_as_ind_other = TRUE,
  show_levels = TRUE
)
}
\arguments{
\item{target_variable}{A character argument indicating for which variable to specify the weighting target. The default "options" will just print the possible variables you can choose.}

\item{.moving_average}{Numeric, whole numbers only, 1 by default. Indicating how many of the past quarters to average over when getting weights for partyid, with proportions provided from Gallup.}

\item{.gallup_url}{The input URL needed to get to the quarterly ratings - it is necessary to contact Gallup to get the URL.}

\item{.stop_date}{Defaults to NULL - present date. Character string of the form YYYY-MM-DD: the date at which to stop the partyid collection, if e.g., you want to weight for data from December 2023 then 2023-12-01.}

\item{partyid_as_ind_other}{If TRUE, partyid will be percentages based on collapsing other parties and don't know responses into Independent. If FALSE, it will work with only Democrat, Independent, and Other specifically and you must code all other responses as NA.}

\item{show_levels}{If TRUE (the default), print the expected levels of the variable to be weighted. Your levels in your data must match these or later weighting stages will not work. If FALSE, it will not print these levels.}
}
\description{
Generate a weighting target (technically, a weights::wpct() object) for a specified variable in the US population. Run the function without any specification to see your options. This function is updated with ACS5 2022 information on the US population.
}
