% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_us_voter_24.R
\name{gen_us_voter_24}
\alias{gen_us_voter_24}
\title{gen_us_voter_24}
\usage{
gen_us_voter_24(
  target_variable = "options",
  partyid_as_ind_other = TRUE,
  show_levels = TRUE
)
}
\arguments{
\item{target_variable}{A character argument indicating for which variable to specify the weighting target. The default "options" will just print the possible variables you can choose.}

\item{partyid_as_ind_other}{If TRUE, partyid will be percentages based on collapsing other parties and don't know responses into Independent. If FALSE, it will work with only Democrat, Independent, and Republican specifically and you must code all other responses as NA.}

\item{show_levels}{If TRUE (the default), print the expected levels of the variable to be weighted. Your levels in your data must match these or later weighting stages will not work. If FALSE, it will not print these levels.}
}
\description{
Generate a weighting target (technically, a weights::wpct() object) for a specified variable in the US population of registered voters. This is based on weighting from the 2024 CCES. Run the function without any specification to see your options.
}
