% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_priors.R
\name{generate_priors}
\alias{generate_priors}
\title{Generate Prior Code}
\usage{
generate_priors(
  items,
  term = "item",
  specific_prior = "normal(0, 1.5)",
  other_prior = "normal(0, 1)",
  sd_prior = "exponential(2)",
  intercept_prior = "normal(0, 1.5)"
)
}
\arguments{
\item{items}{A vector of item names or identifiers for which specific priors will be created.}

\item{term}{A string that defines the prefix for the coefficients. Default is "item".}

\item{specific_prior}{A string that defines the prior distribution for specific items. Default is "normal(0, 1.5)".}

\item{other_prior}{A string that defines the prior distribution for other, non-specific items. Default is "normal(0, 1)".}

\item{sd_prior}{A string that defines the prior distribution for the standard deviation. Default is "exponential(2)".}

\item{intercept_prior}{A string that defines the prior distribution for the intercept. Default is "normal(0, 1.5)".}
}
\value{
A list of prior objects.
}
\description{
This function generates prior code specification for a given set of items. The function has limited uses currently and needs to be made more general.
}
