% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xblock.R
\name{geom_xblock}
\alias{geom_xblock}
\title{geom_xblock}
\usage{
geom_xblock(
  xmin,
  fill = "grey95",
  alpha = 0.5,
  color = "grey95",
  linetype = "blank",
  size = 0
)
}
\arguments{
\item{xmin}{X position where the block will start - call get_xpos()$start with the variables from nice_x()}

\item{fill}{Color for the fill of the block, defaults to grey95}

\item{alpha}{Numeric value from 0 to 1 to indicate alpha/transparency for the block. Defaults to .5}

\item{color}{Color for the line around the variable, defaults to NULL}

\item{linetype}{Linetype for the geom, defaults to "blank"}

\item{size}{Size of the line around the block. Defaults to 0 - i.e., no line around the block}
}
\description{
Add a geom_rect() that functions as a number or annotation block, allowing easy display of numbers on the side. This should be used in conjunction with nice_x, with number_block set to TRUE.
}
