% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_frame.R
\docType{data}
\name{GeomFrame}
\alias{GeomFrame}
\title{Geom Frame}
\format{
An object of class \code{GeomFrame} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomFrame
}
\description{
A geom that places a 'frame' around a ggplot panel
}
\keyword{datasets}
