% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci}
\alias{get_ci}
\title{Get confidence interval}
\usage{
get_ci(
  variable = "outcome",
  data,
  type = "ordinal",
  level = 0.95,
  decimals = 1
)
}
\arguments{
\item{variable}{Character string, indicating the outcome variable name for which confidence interval will be generated}

\item{data}{Your data}

\item{type}{A string indicating whether to outcome is ordinal (or likert, cumulative), binary (or bernoulli, binomial), or categorical (or multinomial, nominal). Defaults to ordinal}

\item{level}{The level of the confidence interval, defaults to .95}

\item{decimals}{How many decimal places to show in the label. Defaults to 1}
}
\description{
Provides frequentist confidence intervals for ordinal, binary, or categorical outcomes.
}
