% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gallup_old.R
\name{get_gallup_old}
\alias{get_gallup_old}
\title{Get Gallup (OLD)}
\usage{
get_gallup_old(moving_average = 3, custom_url = NULL, stop_date = NULL)
}
\arguments{
\item{moving_average}{Numeric input indicating how many months to average over. Defaults to 3.}

\item{custom_url}{An alternative input URL in the event that the standard URL changes/does not work}

\item{stop_date}{Character string of the form YYYY-MM-DD: the date at which to stop the partyid collection, if e.g., you want to weight for data from December 2023 then 2023-12-01}
}
\description{
Retrieve up to date US political party affiliation numbers from Gallup. NOTE: This is the old version of the function that has been shelved due to Gallup changing their policies and availability of monthly data.
}
