% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdi.R
\name{get_hdi}
\alias{get_hdi}
\title{Get highest density interval}
\usage{
get_hdi(
  variable = "outcome",
  data,
  type = "ordinal",
  level = 0.95,
  decimals = 1,
  point_est = "median",
  prior = NULL
)
}
\arguments{
\item{variable}{Character string, indicating the outcome variable name for which confidence interval will be generated}

\item{data}{Your data}

\item{type}{A string indicating whether to outcome is ordinal (or likert, cumulative), binary (or bernoulli, binomial), or categorical (or multinomial, nominal). Defaults to ordinal}

\item{level}{The level of the HDI, defaults to .95}

\item{decimals}{How many decimal places to show in the label. Defaults to 1}

\item{point_est}{Character string, either median, mean, or mode, to select which type of posterior point estimate to include in the label and as the pct output column}

\item{prior}{Optionally set a more specific prior for the model using the typical brms set_prior function, for example: brms::set_prior("student_t(3, 0, 1.5)", class = "Intercept")}
}
\description{
Provides Bayesian highest density intervals for ordinal, binary, or categorical outcomes.
}
