% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ypos.R
\name{get_ypos}
\alias{get_ypos}
\title{get_ypos}
\usage{
get_ypos(
  data,
  variable,
  nearest = 10,
  break_width = NULL,
  number_block_multiple = 1.2
)
}
\arguments{
\item{data}{Copy from call to nice_y() - The data from which the y variable is to be taken}

\item{variable}{Copy from call to nice_y() - A character string naming the y variable}

\item{nearest}{Copy from call to nice_y() - The nearest number to round up to - defaults to 10}

\item{break_width}{Copy from call to nice_y() - Size of the major break, defaults to 'nearest'}

\item{number_block_multiple}{Copy from call to nice_y() - Proportional expansion of the axis to accommodate the number block - defaults to 1.2}
}
\description{
Get the start and middle points for a number/annotation block.
}
