% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\name{hdp}
\alias{hdp}
\title{Highest Density Point}
\usage{
hdp(data)
}
\arguments{
\item{data}{The data for which the region variable will be made}
}
\description{
Get the highest density point (essentially, the mode) of a distribution of values.
}
