% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_png.R
\name{j_png}
\alias{j_png}
\title{Jimbilben PNG}
\usage{
j_png(
  filename = ".png",
  width = 6,
  height = 6,
  units = "in",
  type = "cairo",
  res = 1200,
  full_path = glue::glue("{getwd()}/pngs/{filename}"),
  ...
)
}
\arguments{
\item{filename}{A string specifying the name of the output PNG file. Defaults to ".png".}

\item{width}{A numeric value specifying the width of the PNG file in the given units. Defaults to 6.}

\item{height}{A numeric value specifying the height of the PNG file in the given units. Defaults to 6.}

\item{units}{A string specifying the units for the width and height parameters. Defaults to "in" (inches).}

\item{type}{A string specifying the type of PNG device to use. Defaults to "cairo".}

\item{res}{A numeric value specifying the resolution of the PNG file in dots per inch (dpi). Defaults to 1200.}

\item{...}{Additional arguments passed to the png() function.}

\item{path}{A string specifying the directory path for saving the PNG file. Defaults to a "pngs" folder within the current working directory.}

\item{directory}{A string specifying the current working directory. Defaults to the result of getwd().}
}
\value{
None. The function saves a PNG file to the specified path with the provided parameters.
}
\description{
This function creates a PNG file with specified dimensions and resolution, and saves it to the designated path.
}
\examples{
# Save a PNG file with default settings
j_png("test_plot.png")

# Save a PNG file with custom dimensions and resolution
j_png("test_plot_large.png", width = 12, height = 12, res = 300)
}
