% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linebreaker.R
\name{linebreaker}
\alias{linebreaker}
\title{linebreaker
Place line breaks after a designated number of characters in a string.}
\usage{
linebreaker(text, max_chars = 100, linebreak = "<br>")
}
\arguments{
\item{text}{The text to put line breaks in}

\item{max_chars}{The number of characters after which to place a line break}

\item{linebreak}{Defaults to <br> - the type of line break to inset. Could also be '\n', for example}
}
\description{
linebreaker
Place line breaks after a designated number of characters in a string.
}
