% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunique.R
\name{lunique}
\alias{lunique}
\title{Lazy unique}
\usage{
lunique(variable, data = ref_data)
}
\arguments{
\item{variable}{the variable to assess unique values of}

\item{data}{What data to draw the unique values from - defaults to ref data}
}
\description{
Get unique values from a variable in a data frame with less typing
}
