% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunique2.R
\name{lunique2}
\alias{lunique2}
\title{Lazy unique 2}
\usage{
lunique2(data = ref_data, variable)
}
\arguments{
\item{data}{What data to draw the unique values from - defaults to ref data}

\item{variable}{the variable to assess unique values of}
}
\description{
Get unique values from a variable in a data frame with less typing, formatted to work in a pipe
}
