% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_alpha.R
\name{make_alpha}
\alias{make_alpha}
\title{Make Alpha}
\usage{
make_alpha(colors = c("#56a3cf", "#cecece", "#f2763c"), alpha = 0.67)
}
\arguments{
\item{colors}{A color or vector of colors from which the gradient will be made}

\item{alpha}{A value or vector of values from 0 to 1, to indicate the level of transparency to add to the colors. Defaults to .67 (slight transparency)}
}
\description{
Generate a vector of transparent colors based on color inputs.
}
