% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_colors.R
\name{make_colors}
\alias{make_colors}
\title{Make Colors}
\usage{
make_colors(
  colors = NULL,
  type = "diverging",
  n = 5,
  alpha = 1,
  append_missing = "none",
  missing_colors = "#D9D9D9"
)
}
\arguments{
\item{colors}{A vector of colors from which the gradient will be made}

\item{type}{"diverging" or "linear" - if you don't specify your own colors, we can give a default diverging or linear gradient.}

\item{n}{The number of different colors to select from the gradient e.g., the number of Likert items you are coloring. Defaults to 5.}

\item{alpha}{A value from 0 to 1, to indicate the level of transparency to add to the colors. Defaults to 1 (no transparency)}

\item{append_missing}{Whether an additional color outside the gradient should be added e.g., for missing values or DK responses. Can take values of "start", "end", or "none". Defaults to "none".}

\item{missing_colors}{A single color or vector of colors that will be placed in for values outside the gradient. Defaults to "D9D9D9" - aka 'grey85'}
}
\description{
Generate a vector of colors selected from a gradient between chosen colors.
}
