% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_colours.R
\name{make_colours}
\alias{make_colours}
\title{Make Colours}
\usage{
make_colours(
  colours = NULL,
  type = "diverging",
  n = 5,
  alpha = 1,
  append_missing = "none",
  missing_colours = "#D9D9D9"
)
}
\arguments{
\item{colours}{A vector of colours from which the gradient will be made}

\item{type}{"diverging" or "linear" - if you don't specify your own colours, we can give a default diverging or linear gradient.}

\item{n}{The number of different colours to select from the gradient e.g., the number of Likert items you are colouring. Defaults to 5.}

\item{alpha}{A value from 0 to 1, to indicate the level of transparency to add to the colours. Defaults to 1 (no transparency)}

\item{append_missing}{Whether an additional colour outside the gradient should be added e.g., for missing values or DK responses. Can take values of "start", "end", or "none". Defaults to "none".}

\item{missing_colours}{A single colour or vector of colours that will be placed in for values outside the gradient. Defaults to "D9D9D9" - aka 'grey85'}
}
\description{
Generate a vector of colours selected from a gradient between chosen colours.
}
