% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rescor_formula.R
\name{make_rescor_formula}
\alias{make_rescor_formula}
\title{Make residual correlation formula}
\usage{
make_rescor_formula(my_names)
}
\arguments{
\item{my_names}{A string vector containing the names of the variables to be correlated in your data set}
}
\description{
Generate the brms formula for making pairwise correlations among a set of variables
}
