% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_beta_poststrat.R
\name{mrp_beta_poststrat}
\alias{mrp_beta_poststrat}
\title{MRP Beta Poststrat}
\usage{
mrp_beta_poststrat(
  current_model_epred,
  current_model_phi,
  subgroups = FALSE,
  outcome = "vaccines",
  interval = 0.95,
  nsim = 50000,
  poststrat_tibble = acs5_2020_poststrat_with_partyid,
  .decimals = 2,
  .phi_decimals = 2,
  .point_est = "mean",
  .remove_lead = FALSE,
  .percentage = FALSE
)
}
\arguments{
\item{current_model_epred}{posterior prediction for the regression model of interest - when making the prediction, specify a thinned epred with specific draws the same as for current_model_phi}

\item{current_model_phi}{posterior samples of the phi parameter for the model, from as_draws_df - specify a thinned epred with specific draws the same as for current_model_epred}

\item{subgroups}{Indicating whether this is a population-level or subgroup prediction}

\item{outcome}{Give a name to the outcome being computed (e.g., Support)}

\item{interval}{The summary interval level}

\item{nsim}{How many individual responses to simulate for generating the estimated beta distribution. Defaults to 50000}

\item{poststrat_tibble}{This is the tibble - usually an ACS tibble - containing the demographic variable names. You can group it}

\item{.decimals}{How many decimals for the label that is returned - defaults to 2}

\item{.phi_decimals}{How many decimals for the phi parameter that is returned - defaults to 2}

\item{.point_est}{What point estimate to use for the label summarising the posterior}

\item{.remove_lead}{Whether or not to remove leading zeroes in the label - defaults to FALSE}

\item{.percentage}{Whether to represent the beta proportions as a percentage rather than from 0-1 (defaults to FALSE)}
}
\description{
Post-stratify for beta distributed models. Provides different parameter summaries for the distribution - the mean, the median, and the mode, as these can vary a lot with a beta distribution depending on the level of precision.
}
