% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_normal_poststrat.R
\name{mrp_normal_poststrat}
\alias{mrp_normal_poststrat}
\title{MRP Normal Poststrat}
\usage{
mrp_normal_poststrat(
  current_model_epred,
  subgroups = FALSE,
  outcome = "vaccines",
  interval = 0.95,
  poststrat_tibble = acs5_2020_poststrat_with_partyid,
  .decimals = 2,
  .point_est = "mean"
)
}
\arguments{
\item{current_model_epred}{posterior prediction for the regression model of interest}

\item{subgroups}{Indicating whether this is a population-level or subgroup prediction}

\item{outcome}{Give a name to the outcome being computed (e.g., Support)}

\item{interval}{The summary interval level}

\item{poststrat_tibble}{This is the tibble - usually an ACS tibble - containing the demographic variable names. You can group it}
}
\description{
Post-stratify for normally distributed/gaussian models.
}
