% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_party_ordinal_poststrat.R
\name{mrp_party_ordinal_poststrat}
\alias{mrp_party_ordinal_poststrat}
\title{MRP Party Ordinal Poststrat}
\usage{
mrp_party_ordinal_poststrat(
  current_model_epred,
  subgroups = FALSE,
  outcome = "Vaccine support",
  interval = 0.95,
  poststrat_tibble = acs5_2020_poststrat_with_partyid,
  poststrat_epred = acs5_2020_model_expected_n,
  .point_est = "median",
  .decimals = 1,
  .remove_lead = FALSE
)
}
\arguments{
\item{current_model_epred}{posterior prediction for the regression model of interest}

\item{subgroups}{Indicating whether this is a population-level or subgroup prediction}

\item{outcome}{Give a name to the outcome being computed (e.g., Support)}

\item{interval}{The summary interval level}

\item{poststrat_tibble}{This is the tibble - usually an ACS tibble - containing the demographic variable names. You can group it}

\item{poststrat_epred}{This is the epred for our party poststrat model containing numbers of people expected to fall into each row of the poststrat tibble}
}
\description{
Post-stratify for ordinal models.
}
