% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_party_psup_poststrat.R
\name{mrp_party_psup_poststrat}
\alias{mrp_party_psup_poststrat}
\title{MRP Party PSup Poststrat}
\usage{
mrp_party_psup_poststrat(
  current_model_epred,
  subgroups = FALSE,
  comparison = "e.g., vax_min_treat",
  interval = 0.95,
  superior_name = "Sup",
  inferior_name = "Inf",
  equal_name = "Same",
  poststrat_tibble = acs5_2020_poststrat_with_partyid,
  poststrat_epred = acs5_2020_model_expected_n,
  .point_est = "mean",
  .decimals = 2,
  .remove_lead = TRUE
)
}
\arguments{
\item{current_model_epred}{posterior prediction for the regression model of interest}

\item{subgroups}{Indicating whether this is a population-level or subgroup prediction}

\item{comparison}{Give a name to the comparison being computed (e.g., Vaccines minus Treatments)}

\item{interval}{The summary interval level}

\item{superior_name}{Exact name of the outcome indicating superiority - e.g., "Sup"}

\item{inferior_name}{Exact name of the outcome indicating inferiority - e.g., "Inf"}

\item{equal_name}{Exact name of the outcome indicating equality - e.g., "Same"}

\item{poststrat_tibble}{This is the tibble - usually an ACS tibble - containing the demographic variable names. You can group it}

\item{poststrat_epred}{This is the epred for our party poststrat model containing numbers of people expected to fall into each row of the poststrat tibble}
}
\description{
Post-stratify for within subjects probability of superiority.
}
