% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_age.R
\name{mutate_age}
\alias{mutate_age}
\title{Mutate Age}
\usage{
mutate_age(data, var_name = age_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The target of the mutate - i.e., the variable name in which the numeric ages are stored.}
}
\description{
Mutate a numeric age variable to an age bracket used in MRP or weighting. The new variable will be called 'age'
}
