% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_approval.R
\name{mutate_approval}
\alias{mutate_approval}
\title{Mutate Approval}
\usage{
mutate_approval(data, var_name = pres_approve_nonbin)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The variable in which the Presidential approval categories are contained.}
}
\description{
Mutate a fine-grained presidential approval variable into Approve, Disapprove, and Don't Know. The mutated variable will be called 'pres_approval'.
}
