% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_bible.R
\name{mutate_bible}
\alias{mutate_bible}
\title{Mutate Bible}
\usage{
mutate_bible(data, var_name = bible_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The target of the mutate where the original data is.}
}
\description{
Mutate a Bible views variable with shortened response names used in weighting. The mutated variable will be called 'bible'.
}
