% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_comparison_within.R
\name{mutate_comparison_within}
\alias{mutate_comparison_within}
\title{Mutate Within Subjects Comparison}
\usage{
mutate_comparison_within(
  df,
  a,
  b,
  sep = "_min_",
  append = "psup_",
  psup_convert = TRUE
)
}
\arguments{
\item{df}{The data frame for which the mutate will be made.}

\item{a}{A character vector containing the names of the first set of columns.}

\item{b}{A character vector containing the names of the second set of columns. 'a' and 'b' must have the same length.}

\item{sep}{A character string specifying the separator to be used between the names of 'a' and 'b' in the new column names - defaults to "_min_".}

\item{append}{A character string to be appended before the new column names - defaults to "psup_".}

\item{psup_convert}{A logical indicating whether to convert the difference values into a psup format. If TRUE, negative difference values are converted to 0, positive difference values are converted to 1, and zero difference values remain as 0.5. Defaults to TRUE.}
}
\value{
A dataframe with the original columns and the newly generated columns.
}
\description{
This function takes a data frame and two sets of column names (a and b) and generates new columns that are the differences between the corresponding columns in a and b. The new columns are named by appending 'a' and 'b' with a separator. If psup_convert is set to TRUE, it then converts these difference values into a psup (probability superiority) format.
}
