% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_divisions.R
\name{mutate_divisions}
\alias{mutate_divisions}
\title{Mutate Divisions}
\usage{
mutate_divisions(data, var_name = state)
}
\arguments{
\item{data}{The data for which the division variable will be made}
}
\description{
Assign states to their respective Census divisions with a mutate. The new variable is called 'division'. Use function set_census_divisions() first.
}
