% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_education.R
\name{mutate_education}
\alias{mutate_education}
\title{Mutate Education}
\usage{
mutate_education(data, var_name = education_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The variable in which the education categories are contained.}
}
\description{
Mutate a fine-grained education variable into the larger categories we use for MPR. The mutated variable will be called 'education'.
}
