% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_education_collapse.R
\name{mutate_education_collapse}
\alias{mutate_education_collapse}
\title{Mutate Education Collapse}
\usage{
mutate_education_collapse(data, var_name = education_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The variable in which the education categories are contained.}
}
\description{
Mutate a fine-grained education variable into the larger categories we use for MPR. High school and less than high school are combined. The mutated variable will be called 'education_collapse'.
}
