% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_intfreq.R
\name{mutate_intfreq}
\alias{mutate_intfreq}
\title{Mutate Internet Frequency}
\usage{
mutate_intfreq(data, var_name = intfreq_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The target of the mutate where the original data is.}
}
\description{
Mutate an Internet Frequency variable with 4 collapsed response options used in weighting. The mutated variable will be called 'intfreq_collapse'. Source is Pew https://www.pewresearch.org/wp-content/uploads/2021/03/Internet-Frequency-Update-Methodology-Topline.pdf
}
