% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_intfreq_collapse.R
\name{mutate_intfreq_collapse}
\alias{mutate_intfreq_collapse}
\title{Mutate Internet Frequency Collapse}
\usage{
mutate_intfreq_collapse(data, var_name = intfreq_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The target of the mutate where the original data is.}
}
\description{
Mutate an Internet Frequency variable with 3 collapsed response options used in weighting. The mutated variable will be called 'intfreq_collapse'.
}
