% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_lib_con.R
\name{mutate_lib_con}
\alias{mutate_lib_con}
\title{Mutate Political Views (Liberal - Conservative)}
\usage{
mutate_lib_con(data, var_name = lib_con_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The target of the mutate where the original data is.}
}
\description{
Mutate a Liberal / Conservative Political Views variable with shortened response names used in weighting. The mutated variable will be called 'bible'.
}
