% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_male.R
\name{mutate_male}
\alias{mutate_male}
\title{Mutate Male}
\usage{
mutate_male(data, var_name = sex_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The variable in which the sexual identity categories are contained.}
}
\description{
Mutate a sexual identification variable into a numeric female vs. male variable, with 0 being between them for a regression model - as in MRP. The mutated variable will be called 'male', with -.5 being female, and .5 being male.
}
