% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_partyid.R
\name{mutate_partyid}
\alias{mutate_partyid}
\title{Mutate Partyid}
\usage{
mutate_partyid(data, var_name = party_raw)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{var_name}{The target of the mutate - i.e., the original party ID variable.}
}
\description{
Mutate a political party variable that includes Other and Not Sure into a simplified version where these become Independent. The mutated variable will be called 'partyid'.
}
