% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_psup_label.R
\name{mutate_psup_label}
\alias{mutate_psup_label}
\title{Mutate PSup Label}
\usage{
mutate_psup_label(
  data,
  point_est = "mean",
  lower = "lower",
  upper = "upper",
  size = 5.5,
  linebreak = TRUE,
  subcolor = "#757575"
)
}
\arguments{
\item{data}{The data for which the mutate will be made.}

\item{point_est}{The column name for point estimates - defaults to "mean".}

\item{lower}{The column name for lower bounds of confidence intervals - defaults to "lower".}

\item{upper}{The column name for upper bounds of confidence intervals - defaults to "upper".}

\item{size}{The font size for the confidence interval part of the label - defaults to 5.5.}

\item{linebreak}{A logical indicating whether to put the confidence interval on a new line. If TRUE, a line break is added before the confidence interval. Defaults to TRUE.}
}
\value{
A dataframe with the original columns replaced by a new "perc_label" column.
}
\description{
Converts numeric point estimate and confidence interval columns into a formatted PSup label. If linebreak is set to TRUE, it adds a line break before the confidence interval. If linebreak is set to FALSE, the entire label is on one line.
}
