% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_race.R
\name{mutate_race}
\alias{mutate_race}
\title{Mutate Race}
\usage{
mutate_race(data, race_var_name = race_og, hisp_var_name = hispanic_og)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{race_var_name}{The variable in which racial identities are stored.}

\item{hisp_var_name}{The in which hispanic identification is stored.}
}
\description{
Mutate a variables related to racial/ethnic identity into the categories of racial identification that we use in MRP. The mutated variable will be called 'race'.
}
