% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_race_collapse.R, R/race_collapse.R
\name{mutate_race_collapse}
\alias{mutate_race_collapse}
\title{Mutate Race Collapse}
\usage{
mutate_race_collapse(
  data,
  race_var_name = race_og,
  hisp_var_name = hispanic_og
)

mutate_race_collapse(
  data,
  race_var_name = race_og,
  hisp_var_name = hispanic_og
)
}
\arguments{
\item{data}{The data for which the mutate will be made}

\item{race_var_name}{The variable in which racial identities are stored.}

\item{hisp_var_name}{The in which hispanic identification is stored.}
}
\description{
Mutate a variables related to racial/ethnic identity into the categories of racial identification that we use in MRP and weighting. In this case, Asian/Asian American is categorised along with other. The mutated variable will be called 'race_collapse'.

Mutate a variables related to racial/ethnic identity into the categories of racial identification that we use in MRP and weighting. In this case, Asian/Asian American is categorised along with other. The mutated variable will be called 'race_collapse'.
}
