% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_regions.R
\name{mutate_regions}
\alias{mutate_regions}
\title{Mutate Regions}
\usage{
mutate_regions(data, var_name = state)
}
\arguments{
\item{data}{The data for which the region variable will be made}
}
\description{
Assign states to their respective Census regions with a mutate. The new variable is called 'region'. Use function set_census_regions() first.
}
