% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_number.R
\name{nice_num}
\alias{nice_num}
\title{Nice Number}
\usage{
nice_num(number, decimals = 2, remove_lead = TRUE)
}
\arguments{
\item{number}{The number to be converted to a nice number}

\item{decimals}{The number of decimals to be included, defaults to 2}

\item{remove_lead}{Whether the leading 0 should be removed, defaults to TRUE}
}
\description{
Generate a character version of a number in a nice format for plotting.
}
