% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_x.R
\name{nice_x}
\alias{nice_x}
\title{nice_x}
\usage{
nice_x(
  data,
  variable,
  nearest = 10,
  break_width = NULL,
  number_block_multiple = 1.2,
  number_block = FALSE,
  lower_limit = 0,
  expand_low = 0,
  expand_high = 0
)
}
\arguments{
\item{data}{The data from which the x variable is to be taken}

\item{variable}{A character string naming the x variable}

\item{nearest}{The nearest number to round up to - defaults to 10}

\item{break_width}{Size of the major break, defaults to 'nearest'}

\item{number_block_multiple}{Proportional expansion of the axis to accommodate the number block - defaults to 1.2}

\item{number_block}{Whether to expand the upper limit to include space for an annotation rectangle, defaults to FALSE}

\item{lower_limit}{The lower limit of the x axis, defaults to 0}

\item{expand_low}{Additive expansion for the low end of x, defaults to 0}

\item{expand_high}{Additive expansion for the high end of x, defaults to 0}
}
\description{
Generate a nicely functioning x axis based on the numbers of the x variable. Includes possible expansion for a number/annotation block.
}
