% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_mrp.R
\name{numeric_mrp}
\alias{numeric_mrp}
\title{Run Numeric MRP}
\usage{
numeric_mrp(
  variable_name,
  variable_label = NULL,
  center_value = set_center_value,
  my_data = set_my_data,
  save_model = save_my_model,
  save_epred = save_my_epred,
  my_prior = c(set_prior("normal(0 , 2)", class = "Intercept"),
    set_prior("normal(0 , 1)", class = "b"), set_prior("exponential(2)", class = "sd"),
    set_prior("normal(1, 1)", class = "sigma", lb = 0.01)),
  my_init = 0,
  my_refresh = 250,
  my_iter = set_my_iter,
  my_warmup = set_my_warmup,
  my_poststrat = set_my_poststrat,
  my_adapt_delta = set_my_adapt_delta,
  mrp_form = NULL,
  name_addition = ""
)
}
\arguments{
\item{variable_name}{String. The name of the numeric variable to be modeled.}

\item{variable_label}{Optional. A human-readable label for the variable. Defaults to \code{variable_name}.}

\item{center_value}{Numeric. A value to center the variable (e.g., 4 to make 4/neutral become 0). Defaults to \code{set_center_value}.}

\item{my_data}{A data frame containing the data for the regression model. Defaults to \code{set_my_data}.}

\item{save_model}{Logical. If \code{TRUE}, saves the fitted model object to an mrp_models folder in your directory. Defaults to \code{save_my_model}.}

\item{save_epred}{Logical. If \code{TRUE}, saves the posterior epred draws to an mrp_epreds folder in your directory. Defaults to \code{save_my_epred}.}

\item{my_prior}{A list of prior distributions for the model. Defaults to:
\itemize{
  \item \code{Intercept: normal(0, 2)}
  \item \code{Coefficients: normal(0, 1)}
  \item \code{Random effects: exponential(2)}
  \item \code{Residual standard deviation: normal(1,1), lower bound 0.01}
}}

\item{my_init}{Initial values for the model. Default is \code{0}.}

\item{my_refresh}{Frequency of progress updates during model fitting. Default is \code{250}.}

\item{my_iter}{Total number of iterations per chain. Default is \code{set_my_iter}.}

\item{my_warmup}{Number of warmup iterations for the sampler. Default is \code{set_my_warmup}.}

\item{my_poststrat}{A data frame for poststratification. Default is \code{set_my_poststrat}.}

\item{my_adapt_delta}{Numeric value for the Stan sampler control parameter to improve convergence. Default is \code{set_my_adapt_delta}.}

\item{mrp_form}{Defaults to NULL, in which case our standard MRP formula is used, but can be changed to reflect a different MRP formula}
}
\value{
A list containing:
  \itemize{
    \item \code{"model"}: The fitted \code{brms} model object.
    \item \code{"epred"}: Posterior predictions from the fitted model.
  }
}
\description{
Perform the regression and epred steps of a multilevel regression and poststratification (MRP) for a numeric outcome variable.
}
\details{
This function performs multilevel Bayesian regression using the \code{brms} package.
It constructs a model formula based on a predefined structure, fits the model
using the \code{cmdstanr} backend, and generates posterior epred draws for
a target population. The model and predictions can be optionally saved for future use.

If the variable is ordinal or categorical, it will be converted to numeric.
Ensure that the ordering of factors is correctly set using \code{ordered = TRUE}.
}
\note{
Requires the \code{brms} and \code{cmdstanr} packages.
Ensure that the \code{my_data} and \code{my_poststrat} arguments are properly formatted.
}
