% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_code.R
\name{percent_code}
\alias{percent_code}
\title{Percent code}
\usage{
percent_code(axis = "x")
}
\arguments{
\item{axis}{String, the axis being referenced - defaults to "x" (vs. "y")}
}
\description{
Print out code for ggplot axis expansion
}
