% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_within.R
\name{permute_within}
\alias{permute_within}
\title{Permute Within Subjects Comparison}
\usage{
permute_within(
  data,
  a,
  b,
  direction = "two-sided",
  binary = FALSE,
  sims = 2500,
  current_prefix = "item_",
  new_prefix = "comp_",
  sep = "_vs_",
  verbose = FALSE
)
}
\arguments{
\item{data}{The data frame for which the permutation test will be conducted.}

\item{a}{The name of the first column in the pair for permutation testing.}

\item{b}{The name of the second column in the pair for permutation testing.}

\item{direction}{Specifies the expected direction of the difference between the pair of columns. It can be "superior" (a > b), "inferior" (a < b), or "two-sided" (non-directional). Defaults to "two-sided".}

\item{binary}{A logical indicating whether the outcome is a binary outcome (in the form of 0 and 1) - if TRUE a permutation test on the ratio of the pairs will also be conducted. Defaults to FALSE.}

\item{sims}{The number of random permutations to perform. Defaults to 2500.}

\item{current_prefix}{The current prefix of the column names in the data frame. Defaults to "item_".}

\item{new_prefix}{The new prefix to be appended before the new column names. Defaults to "comp_".}

\item{sep}{A character string specifying the separator to be used between the names of 'a' and 'b' in the new column name. Defaults to "_vs_".}

\item{verbose}{A logical indicating whether to print detailed output. If TRUE, it returns both the detailed comparisons for each simulation and the summary. If FALSE, it returns only the summary. Defaults to FALSE.}
}
\value{
A dataframe containing the summary of the permutation tests. If verbose = TRUE, a list containing both the detailed comparisons for each simulation and the summary is returned.
}
\description{
This function conducts permutation testing for paired (i.e., within subjects) ratings. It will provide estimates for 'Probability of Superiority' and 'Observation oriented modelling', in which tied values count against any directional difference. If Binary is TRUE, it additionally permutes a ratio.
}
