% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststrat_multi_beta.R
\name{poststrat_multi_beta}
\alias{poststrat_multi_beta}
\title{Poststratify Beta MRP Predictions}
\usage{
poststrat_multi_beta(
  input,
  variable_name,
  outcome_name = NULL,
  subgroups = c("education_collapse", "race", "income_ces", "male", "partyid",
    "age_fine", "region"),
  subgroup_mapping = list(education_collapse = "Education", race = "Race", income_ces =
    "Income", male = "Sex", partyid = "Party affiliation", age_fine = "Age", region =
    "Region"),
  save_output = set_save_output,
  .nsim = 1e+05,
  return_state = set_state,
  .poststrat_tibble = set_my_poststrat,
  .poststrat_epred = set_my_epred
)
}
\arguments{
\item{input}{The object generated from a Beta MRP model, containing the posterior predictions and precision parameter (\code{phi}).}

\item{variable_name}{String. The name of the variable being poststratified.}

\item{outcome_name}{Optional. A human-readable name for the outcome. Defaults to \code{variable_name}.}

\item{subgroups}{A character vector of subgroup variables for stratification. Defaults to:
\itemize{
  \item \code{"education_collapse"}
  \item \code{"race"}
  \item \code{"income_ces"}
  \item \code{"male"}
  \item \code{"partyid"}
  \item \code{"age_fine"}
  \item \code{"region"}
}}

\item{subgroup_mapping}{A named list mapping subgroup variable names to human-readable labels. Defaults to:
\itemize{
  \item \code{"education_collapse"} = \code{"Education"}
  \item \code{"race"} = \code{"Race"}
  \item \code{"income_ces"} = \code{"Income"}
  \item \code{"male"} = \code{"Sex"}
  \item \code{"partyid"} = \code{"Party affiliation"}
  \item \code{"age_fine"} = \code{"Age"}
  \item \code{"region"} = \code{"Region"}
}}

\item{save_output}{Logical. Whether to save the poststratification summaries as an RDS file. Defaults to \code{set_save_output}.}

\item{.nsim}{Numeric. The number of 'people' from the population to be simulated when generating the mean, median, and mode of the distribution. Defaults to \code{100000}.}

\item{return_state}{Logical. Whether to include the "state" subgroup in the output. Defaults to \code{set_state}.}

\item{.poststrat_tibble}{A data frame containing the poststratification tibble. Defaults to \code{set_my_poststrat}.}

\item{.poststrat_epred}{A data frame containing the poststratification posterior predictions. Defaults to \code{set_my_epred}.}
}
\value{
A list containing poststratification summaries:
  \itemize{
    \item \code{"population"}: Population-level summary.
    \item Subgroup summaries for each variable in \code{subgroups}.
  }
}
\description{
Perform poststratification on Beta regression model predictions across population and specified subgroups.
}
