% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psup_steps_code.R
\name{psup_steps_code}
\alias{psup_steps_code}
\title{PSup steps code}
\usage{
psup_steps_code(item_prefix = "item")
}
\description{
Print out code for the steps to conduct Bayesian within subjects PSup
}
