% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_scatter.R
\name{quick_scatter}
\alias{quick_scatter}
\title{Quick Scatter Plot for Exploratory Analysis}
\usage{
quick_scatter(
  x,
  y,
  plot_data = ai_data,
  corr = TRUE,
  fitline = c("lm", "loess"),
  descriptives = TRUE,
  df = 3,
  jitter_height = 0.2,
  jitter_width = 0.2,
  alpha = 0.2,
  dot_color = "grey65",
  ...
)
}
\arguments{
\item{x}{A string representing the name of the x-axis variable}

\item{y}{A string representing the name of the y-axis variable}

\item{plot_data}{The data frame containing the x and y variables (default: ai_data)}

\item{corr}{Logical, whether to display the Pearson and Spearman correlation coefficients (default: TRUE)}

\item{fitline}{A character vector specifying the type(s) of fitline to display: "lm", "spline", or "loess" (default: c("lm", "loess"))}

\item{descriptives}{Logical, whether to display summary statistics on the plot (default: TRUE)}

\item{df}{Degrees of freedom for the spline fitline (default: 3)}

\item{jitter_height}{Numeric, the height of the jitter in geom_jitter (default: .2)}

\item{jitter_width}{Numeric, the width of the jitter in geom_jitter (default: .2)}

\item{alpha}{Numeric, the transparency of the jitter points (default: .2)}

\item{dot_color}{A string representing the color of the geom_jitter points (default: "grey65")}

\item{...}{Additional arguments passed to ggplot2::aes()}
}
\value{
A ggplot2 scatter plot object
}
\description{
This function creates a scatter plot to explore the relationship between two numeric variables,
with options to display a correlation, linear regression, spline, or loess fitline,
and display summary statistics on the plot.
}
\examples{
quick_scatter(x = "income", y = "happiness", plot_data = my_data, fitline = c("lm", "loess"), dot_color = "#336699")

}
