% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{Read Qualtrics}
\usage{
read_qualtrics(
  file,
  provider = "Prolific",
  row_removal = TRUE,
  preview_removal = TRUE,
  convert_time = TRUE,
  extras_removal = TRUE
)
}
\arguments{
\item{file}{An xlsx file from qualtrics}

\item{provider}{The survey or panel provider from which the sample comes, defaults to Prolific}

\item{row_removal}{Should the second row - usually just bad column names - be removed, defaults to TRUE}

\item{preview_removal}{Should preview responses be removed, defaults to TRUE}

\item{convert_time}{Should timer-question-based responses be converted to a number, defaults to TRUE}

\item{extras_removal}{Should empty columns such as recipient name and email be removed, defaults to TRUE}
}
\description{
Reads in a qualtrics excel file. The redundant second row is removed, as are preview responses. Columns identified with time are converted to numbers.
}
