% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_numbers.R
\name{reverse_numbers}
\alias{reverse_numbers}
\title{Reverse numbers}
\usage{
reverse_numbers(numbers, minmax = NULL)
}
\arguments{
\item{numbers}{A numeric vector to be reversed.}

\item{minmax}{A numeric vector for the minimum and maximum values - defaults to NULL, in which case the observed min and max from the data are used}
}
\value{
A numeric vector of the same length as the input, with values reversed within the range of the input vector.
}
\description{
This function reverses the numeric values within a vector by subtracting each number
from the maximum value and then adding the minimum value.
}
\examples{
my_numbers <- c(2, 4, 6, 8, 10)
reverse_numbers(my_numbers)
}
