% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_down.R
\name{round_down}
\alias{round_down}
\title{Round down}
\usage{
round_down(number, accuracy = 10, f = floor)
}
\arguments{
\item{number}{The number to be rounded up}

\item{accuracy}{What to round up to the nearest multiple of}
}
\description{
Round a number down to the nearest specified number
}
