% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{Round up}
\usage{
round_up(number, accuracy, f = ceiling)
}
\arguments{
\item{number}{The number to be rounded up}

\item{accuracy}{What to round up to the nearest multiple of}
}
\description{
Round a number up to the nearest specified number
}
