% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_census_regions.R
\name{set_census_regions}
\alias{set_census_regions}
\title{Set Census Regions}
\usage{
set_census_regions(set = TRUE, print = FALSE)
}
\arguments{
\item{set}{If TRUE, this will define the region variables to your R environment, set to TRUE by default}

\item{print}{If TRUE, this will just print out the region definitions to your R console, set to TRUE by default}
}
\description{
Set variables in your R environment representing different regions of the USA from the Census https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf and the states that comprise them. This is necessary for weighting and MRP.
}
