% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colors.R
\name{set_colors}
\alias{set_colors}
\title{Set colors}
\usage{
set_colors(print = TRUE)
}
\arguments{
\item{print}{Logical, whether or not to print the names of the colors that have been set}
}
\description{
Set a range of colors as objects in the environment
}
