% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_custom_regions.R
\name{set_custom_regions}
\alias{set_custom_regions}
\title{Set Custom Regions}
\usage{
set_custom_regions(set = TRUE, print = FALSE)
}
\arguments{
\item{set}{If TRUE, this will define the region variables to your R environment, set to TRUE by default}

\item{print}{If TRUE, this will just print out the region definitions to your R console, set to TRUE by default}
}
\description{
Set variables in your R environment representing different regions of the USA and the states that comprise them. This is necessary for weighting and MRP.
}
