% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_data.R
\name{set_ref_data}
\alias{set_ref_data}
\title{Set Reference Data}
\usage{
set_ref_data(data, name = NULL)
}
\arguments{
\item{data}{The data to be set as the reference data}

\item{name}{A string - if you want the reference data to be named something other than ref_data, provide a name here}
}
\description{
Set reference data to be referred to by some other functions such as lunique()
}
