% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squish_ratings.R
\name{squish_ratings}
\alias{squish_ratings}
\title{Squish Ratings}
\usage{
squish_ratings(ratings, max_rating = NULL, proportion = 0.2)
}
\arguments{
\item{ratings}{A numeric vector of ratings to be shifted.}

\item{max_rating}{The maximum rating value. If NULL (default), the maximum value of the input `ratings` vector will be used.}

\item{proportion}{A numeric value (between 0 and 1) specifying the proportion by which to shift the ratings (default is 0.2).}
}
\value{
A numeric vector of shifted ratings.
}
\description{
This function squishes a vector of ratings over based on the maximum rating value and a specified shift proportion.
}
\examples{
ratings <- c(30, 50, 90)
max_rating <- 100
proportion <- 0.1

squished_ratings <- squish_ratings(ratings, max_rating, proportion)
print(squished_ratings)

}
