% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_code.R
\name{summarise_code}
\alias{summarise_code}
\title{Print summarise code}
\usage{
summarise_code(outcome = "rating")
}
\arguments{
\item{outcome}{The name of the variable that is being summarised}
}
\description{
Print out code for a commonly desired summary and corresponding label
}
