% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_binom_confint.R
\name{svy_binom_confint}
\alias{svy_binom_confint}
\title{Survey Binomial Confidence Interval}
\usage{
svy_binom_confint(
  object,
  interval = 0.95,
  decimals = 1,
  percentage = TRUE,
  remove_lead = TRUE,
  provide_inverse = TRUE,
  present = 1,
  absent = 0
)
}
\arguments{
\item{object}{A survey weighted binomial regression fit model object.}

\item{interval}{The width of the confidence interval to be calculated, defaults to .95.}

\item{decimals}{The number of decimals to be included in summary labels, defaults to 1.}

\item{percentage}{Whether to represent the estimates as percentages, defaults to TRUE.}

\item{remove_lead}{Whether the leading 0 should be removed in summary labels, defaults to TRUE.}

\item{provide_inverse}{Whether to provide the inverse probability estimates, defaults to TRUE.}

\item{present}{The value representing presence, defaults to 1.}

\item{absent}{The value representing absence, defaults to 0.}
}
\value{
A tibble with columns for the estimate, lower and upper bounds of the confidence interval, and a formatted label. If `provide_inverse` is TRUE, includes the inverse probability estimates.
}
\description{
Generate a tibble summarizing the estimates and confidence intervals of a survey weighted binomial regression fit model.
}
