% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_cat_confint.R
\name{svy_cat_confint}
\alias{svy_cat_confint}
\title{Survey Categorical Confidence Interval}
\usage{
svy_cat_confint(
  object,
  data,
  category_name = "outcome",
  interval = 0.95,
  decimals = 1,
  percentage = TRUE,
  remove_lead = TRUE
)
}
\arguments{
\item{object}{A survey weighted categorical regression fit model object.}

\item{data}{The data frame used to fit the model.}

\item{category_name}{The name of the categorical outcome variable, defaults to "outcome".}

\item{interval}{The width of the confidence interval to be calculated, defaults to .95.}

\item{decimals}{The number of decimals to be included in summary labels, defaults to 1.}

\item{percentage}{Whether to represent the estimates as percentages, defaults to TRUE.}

\item{remove_lead}{Whether the leading 0 should be removed in summary labels, defaults to TRUE.}
}
\value{
A tibble with columns for the estimate, lower and upper bounds of the confidence interval, standard error, and formatted labels.
}
\description{
Generate a tibble summarizing the estimates and confidence intervals of a survey weighted categorical regression fit model.
}
