% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_normal_confint.R
\name{svy_normal_confint}
\alias{svy_normal_confint}
\title{Survey Normal Confidence Interval}
\usage{
svy_normal_confint(object, interval = 0.95, decimals = 1, remove_lead = TRUE)
}
\arguments{
\item{object}{A survey weighted regression fit model object.}

\item{interval}{The width of the confidence interval to be calculated, defaults to .95.}

\item{decimals}{The number of decimals to be included in summary labels, defaults to 1.}

\item{remove_lead}{Whether the leading 0 should be removed in summary labels, defaults to TRUE.}
}
\value{
A tibble with columns for the estimate, lower and upper bounds of the confidence interval, and a formatted label.
}
\description{
Generate a tibble summarizing the estimates and confidence intervals of a survey weighted regression fit model.
}
