% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_extract.R
\name{tag_extract}
\alias{tag_extract}
\title{Extract text from specific XML/HTML tags}
\usage{
tag_extract(text, tag = "response")
}
\arguments{
\item{text}{The full text string from which to extract the tagged content.}

\item{tag}{The name of the tag from which to extract the content, without angle brackets. Defaults to "response".}
}
\value{
A character string containing the text extracted from between the specified tags.
If the tag does not exist in the text, the function will return the original text.
}
\description{
This function extracts text enclosed within specified XML or HTML tags from a given text string.
The function uses a regular expression to match and extract the content. It requires the text
from which to extract the content and the specific tag name to look for.
}
\examples{
text_example <- "<customtag>Here is the content</customtag>"
tag_extract(text_example, "customtag")

}
