% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jimbilben_dark.R
\name{theme_jimbilben_dark}
\alias{theme_jimbilben_dark}
\title{Theme Jimbilben Dark}
\usage{
theme_jimbilben_dark(
  base_size = 11,
  legend_vjust = 0.075,
  legend_hjust = 0,
  base_family = "Jost",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  grid_line_color = "white",
  x_ticks = FALSE,
  y_ticks = FALSE,
  top_margin = 5,
  right_margin = 15,
  bottom_margin = 5,
  left_margin = 5
)
}
\arguments{
\item{base_size}{Base size of text - defaults to 12}

\item{legend_vjust}{vjust of the legend title}

\item{legend_hjust}{hjust of the legend title}

\item{base_family}{Overall font family - defaults to Jost}

\item{grid_line_color}{Color for the grid lines - defaults to white}

\item{x_ticks}{TRUE of FALSE, whether to include x ticks or not - defaults to FALSE}

\item{y_ticks}{TRUE of FALSE, whether to include y ticks or not - defaults to FALSE}

\item{top_margin}{top margin, defaults to 5}

\item{right_margin}{top margin, defaults to 15}

\item{bottom_margin}{top margin, defaults to 5}

\item{left_margin}{top margin, defaults to 5}
}
\description{
A good theme with a grey panel background. All text elements are ggtext::element_markdown()
}
