% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_beta.R
\name{transform_beta}
\alias{transform_beta}
\title{Beta Transform Function}
\usage{
beta_transform(number, sample_size = 886, max = 100)
}
\arguments{
\item{number}{The number to be converted.}

\item{sample_size}{An integer representing the sample size for the transformation. Default is 2000.}

\item{max}{An optional numeric value representing the expected maximum value in the data.
All elements in 'number' are divided by this value to initially scale them to a (0,1) range. Default is 100.}
}
\value{
A numeric vector of the same length as 'number', with all elements transformed to fall within the open interval (0,1).
}
\description{
Transforms a number to fall within the open interval (0,1), via a beta transformation. The transformation is especially useful in preparing
data for beta regression models, which require inputs to be in the open interval (0,1).
}
\details{
The function works in two main steps. First, it scales the input vector ('number') to a (0,1) range by dividing by 'max'.
Then it adjusts the scaled values to fall within the open interval (0,1) using the formula:
(number * (sample_size - 1) + 0.5) / sample_size. This ensures that the transformed values are compatible with beta regression requirements.
}
