% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfurl_choices.R
\name{unfurl_choices}
\alias{unfurl_choices}
\title{Unfurl choices
Generate new columns indicating the presence or absence of each possible item in a multiple choice question where people can select more than one thing.}
\usage{
unfurl_choices(
  data,
  variable,
  prefix = "choice_",
  sep = ",",
  choice_options = NULL,
  choice_names = NULL,
  present = 1,
  absent = 0
)
}
\arguments{
\item{data}{the target dataframe}

\item{variable}{the variable/column name in which the multiple choice items are stored}

\item{prefix}{= A string. Each variable name should, but does not have to be, preceded by a prefix so that the new variables can be easily identified/selected and e.g., pivot_longer from. Defaults to choice_}

\item{sep}{Defaults to ','. What separator separates the different choices when more than one is selected?}

\item{choice_options}{If NULL - the default - the unique items will be generated from those that are present in the column. You can also supply your own e.g., to confirm if some known choice was never selected. You can use the unique_choices() function to generate this vector from the data.}

\item{choice_names}{Can only be used in combination with choice_options. A vector of the same length as choice_options indicating the desired variable name that corresponds to each new choice column that will be created.}

\item{present}{Defaults to 1 - what the new variable should be if the item is present. Must be of same type as 'absent'}

\item{absent}{Defaults to 0 - what the new variable should be if the item is absent. Must be of same type as 'present'}
}
\description{
Unfurl choices
Generate new columns indicating the presence or absence of each possible item in a multiple choice question where people can select more than one thing.
}
