% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_ordered_pairs.R
\name{unique_ordered_pairs}
\alias{unique_ordered_pairs}
\title{Unique ordered pairs}
\usage{
unique_ordered_pairs(x)
}
\arguments{
\item{x}{Character vector - the unique items, in the order that they should appear from highest to lowest values}
}
\value{
A tibble with pairs of items
}
\description{
Returns unique pairwise items from a list
}
