% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_pairs.R
\name{unique_pairs}
\alias{unique_pairs}
\title{Unique Pairs}
\usage{
unique_pairs(a = 1:5, b = NULL, sep = " - ")
}
\arguments{
\item{a}{A vector for the first set of items}

\item{b}{A vector for the second set of items - if NULL (the default) then it will just be the same as a}

\item{sep}{A string indicating how to separate the two items when pasting them together. Defaults to" " - "}
}
\description{
Get a list of unique pairs of items, for use in making pairwise comparisons or correlations.
}
