% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cat.R
\name{weighted_cat}
\alias{weighted_cat}
\title{Weighted Categorical Regression}
\usage{
weighted_cat(
  data,
  var_name,
  categ_prior = c(brms::set_prior("normal(0 , 1)", class = "Intercept"))
)
}
\arguments{
\item{data}{A dataframe containing the categorical variable and weights used in the model.}

\item{var_name}{A string specifying the name of the categorical variable of interest.}

\item{categ_prior}{A list of priors for the model's intercepts, typically including normal priors for the intercepts
with mean 0 and standard deviation 1.}
}
\value{
A dataframe summarizing the estimated proportions for each category of the variable,
        neatly formatted with the variable name relocated to the first column and all numeric results rounded
        to one decimal place. Additional formatting modifies the variable name in specific contexts.
}
\description{
This function performs a weighted regression analysis for categorical variables using Bayesian methods.
It estimates the proportions of each category adjusted by weights provided in the data.
}
\details{
The function fits a Bayesian categorical model using `brms::brm` with weighted data to accurately represent the
distribution of categories as influenced by the specified weights. Posterior predictions are extracted using
`tidybayes::add_epred_draws` and summarized using `jimbilben::nice_post` to calculate mean proportions and express
them as percentages.
}
\note{
To successfully run this function, ensure the following packages are installed and loaded: `brms`, `tidybayes`,
      `dplyr`, `tibble`, `glue`, and `stringr` for string manipulations in the final output adjustments.
}
\examples{
weighted_cat(data = my_data, var_name = "my_category_var")

}
