% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_psup.R
\name{weighted_psup}
\alias{weighted_psup}
\title{Weighted Probability of Superiority}
\usage{
weighted_psup(
  data,
  var_name,
  psup_prior = c(brms::set_prior("normal(0 , 1)", class = "Intercept"),
    brms::set_prior("normal(0 , 1)", class = "Intercept", dpar = "mu0"),
    brms::set_prior("normal(0 , 1)", class = "Intercept", dpar = "mu1"))
)
}
\arguments{
\item{data}{A dataframe containing the psup comparison variable and weights used in the model.}

\item{var_name}{A string specifying the name of the psup variable of interest.}

\item{psup_prior}{A list of priors for the model's intercepts, tailored for the categorical levels and weights. By default,
it includes normal priors for three intercepts with mean 0 and standard deviation 1.}
}
\value{
A dataframe summarizing the weighted psup values for the target variable.
}
\description{
This performs a weighted regression to generate probability of superiority.
}
\details{
The function internally fits a Bayesian categorical model with weights using `brms::brm` and converts the values to PSup. Posterior predictions
are performed via `tidybayes::add_epred_draws` to generate expected PSup values. The final output is
processed with `jimbilben::nice_post` to produce a clean summary table.
}
\note{
Ensure that `brms`, `tidybayes`, `dplyr`, `tibble`, and `glue` packages are installed and loaded into your R session
      to use this function without issues.
}
\examples{
weighted_psup(data = my_data, var_name = "my_psup_var")

}
