% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widener.R
\name{widener}
\alias{widener}
\title{Widener}
\usage{
widener(
  for_widening,
  decimals_for_num = decimal_choice,
  remove_lead_for_num = remove_lead_choice
)
}
\arguments{
\item{decimals}{Inherits from the arguments to nice_post}

\item{remove_lead}{Inherits from the arguments to nice_post}
}
\description{
Only used inside of nice_post in order to incorporate multiple HDIs
}
