% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gallup.R, R/get_gallup_old.R
\name{get_gallup}
\alias{get_gallup}
\title{Get Gallup}
\usage{
get_gallup(moving_average = 3, custom_url = NULL, stop_date = NULL)

get_gallup(moving_average = 3, custom_url = NULL, stop_date = NULL)
}
\arguments{
\item{moving_average}{Numeric input indicating how many months to average over. Defaults to 3.}

\item{custom_url}{An alternative input URL in the event that the standard URL changes/does not work}

\item{stop_date}{Character string of the form YYYY-MM-DD: the date at which to stop the partyid collection, if e.g., you want to weight for data from December 2023 then 2023-12-01}

\item{gallup_url}{The input URL needed to get to the quarterly ratings - it is necessary to contact Gallup to get the URL.}
}
\description{
Retrieve up to date US political party affiliation numbers from Gallup. You are now required to request access to quarterly party affiliation data from Gallup, and can then enter the appropriate URL for the data. Usage of the data must involve attribution to Gallup, including when this is just used to weight survey results - see the bottom of the page here: https://news.gallup.com/poll/15370/party-affiliation.aspx

Retrieve up to date US political party affiliation numbers from Gallup. NOTE: This is the old version of the function that has been shelved due to Gallup changing their policies and availability of monthly data.
}
