% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gallup.R
\name{get_gallup}
\alias{get_gallup}
\title{Get Gallup}
\usage{
get_gallup(
  moving_average = 1,
  gallup_url = "you need to request the URL from gallup",
  stop_date = NULL
)
}
\arguments{
\item{moving_average}{Numeric input indicating how many quarters to average over. Defaults to 1.}

\item{gallup_url}{The input URL needed to get to the quarterly ratings - it is necessary to contact Gallup to get the URL.}

\item{stop_date}{Defaults to NULL, i.e., using the most recent available data. Character string of the form 'YYYY QQ': the date at which to stop the partyid collection, if e.g., you want to weight for data from Q2 2023 then '2023 Q2'}
}
\description{
Retrieve up to date US political party affiliation numbers from Gallup. You are now required to request access to quarterly party affiliation data from Gallup, and can then enter the appropriate URL for the data. Usage of the data must involve attribution to Gallup, including when this is just used to weight survey results - see the bottom of the page here: https://news.gallup.com/poll/15370/party-affiliation.aspx
}
