% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gallup.R
\name{get_gallup}
\alias{get_gallup}
\title{Get Gallup}
\usage{
get_gallup(moving_average = 3, custom_url = NULL)
}
\arguments{
\item{moving_average}{Numeric input indicating how many months to average over. Defaults to 3.}

\item{custom_url}{An alternative input URL in the event that the standard URL changes/does not work}
}
\description{
Retrieve uptodate US political party affiliation numbers from Gallup.
}
