% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jimbilben.R
\name{theme_jimbilben}
\alias{theme_jimbilben}
\title{Theme Jimbilben}
\usage{
theme_jimbilben(
  base_size = 12,
  base_family = "Jost",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  border_line_color = "grey95",
  grid_line_color = "grey95",
  x_ticks = FALSE,
  y_ticks = FALSE,
  right_margin = 15
)
}
\arguments{
\item{base_size}{Base size of text - defaults to 12}

\item{base_family}{Overall font family - defaults to Jost}

\item{border_line_color}{Color for the border around the plot - defaults to grey95}

\item{grid_line_color}{Color for the grid lines - defaults to grey95}

\item{x_ticks}{TRUE of FALSE, whether to include x ticks or not - defaults to FALSE}

\item{y_ticks}{TRUE of FALSE, whether to include y ticks or not - defaults to FALSE}
}
\description{
A good theme with a grey panel background. All text elements are ggtext::element_markdown()
}
