% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cortest.R
\name{tidy_cortest}
\alias{tidy_cortest}
\title{Tidy Correlation Test}
\usage{
tidy_cortest(
  x,
  y,
  data,
  method = "spearman",
  alternative = "two.sided",
  threshold = 0.01,
  continuity = TRUE,
  sep = " - ",
  ...
)
}
\arguments{
\item{x}{A character string or variable name representing the first variable for the correlation test.}

\item{y}{A character string or variable name representing the second variable for the correlation test.}

\item{data}{A data frame containing the variables specified in `x` and `y`.}

\item{method}{A character string specifying the correlation method. Options are "spearman" (default), "pearson", and "kendall".}

\item{alternative}{A character string specifying the alternative hypothesis for the correlation test. Options are "two.sided" (default), "less", and "greater".}

\item{threshold}{A numeric value indicating the significance level for flagging correlations as significant. Default is 0.01.}

\item{continuity}{A logical value indicating whether to apply a continuity correction in the correlation test. Default is TRUE.}

\item{sep}{A character string specifying the separator between variable names in the 'pairing' column of the output. Default is " - ".}

\item{...}{Additional arguments passed to `cor.test()`.}
}
\value{
A data frame with the following columns:
  - estimate: The correlation coefficient.
  - p_value: The p-value of the correlation test.
  - statistic: The test statistic.
  - x_var: The name of the first variable.
  - y_var: The name of the second variable.
  - method: The correlation method used (Spearman, Pearson, or Kendall).
  - sig: A logical value indicating whether the correlation is significant based on the specified threshold.
  - pos_neg: The direction of the correlation (Positive, Negative, or Neutral).
  - pairing: A combination of the variable names separated by the specified separator.
}
\description{
This function computes the correlation between two numeric variables and returns a tidy data frame with the results.
It supports Pearson, Spearman, and Kendall correlation methods.
}
\examples{
data(mtcars)
tidy_cortest(mpg, disp, data = mtcars)
tidy_cortest("mpg", "disp", data = mtcars, method = "pearson")
tidy_cortest(mpg, disp, data = mtcars, alternative = "greater", threshold = 0.05)

}
